module for the Parser + register unit and integration tests of the Parser
